refactor(cli): split main.rs into command modules
