+++ crates/dirt-cli/src/commands/mod.rs
@@ -1,0 +1,11 @@
+pub mod add;
+pub mod auth_cmd;
+pub mod common;
+pub mod completions;
+pub mod config;
+pub mod delete;
+pub mod edit;
+pub mod export;
+pub mod list;
+pub mod search;
+pub mod sync;
